+++ src/main/java/chess/domain/piece/Bishop.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.piece;
 
 public class Bishop extends Piece {
