+++ src/main/java/chess/domain/piece/Pawn.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.piece;
 
 public class Pawn extends Piece {
     public Pawn(final Color color) {
