+++ src/main/java/chess/domain/piece/Rook.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.piece;
 
 public class Rook extends Piece {
     public Rook(final Color color) {
