+++ src/test/java/chess/util/TestPiece.java
@@ -1,7 +1,7 @@
 package chess.util;
 
-import chess.domain.Color;
-import chess.domain.Piece;
+import chess.domain.piece.Color;
+import chess.domain.piece.Piece;
 
 public class TestPiece extends Piece {
     public TestPiece(final Color color) {
